CLI help and initd script output.
