+++ .env.example
@@ -1,2 +1,3 @@
 PIFM_PATH=/home/pi/PiFmRds/src/pi_fm_rds
 NODE_ENV=production
+HOST=localhost
