+++ package.json
@@ -5,10 +5,11 @@
   "main": "./lib/encoder.js",
   "dependencies": {
     "mopidy": "^0.4.1",
-    "when": "^3.6.3"
+    "when": "^3.6.3",
+    "yargs": "^1.3.3"
   },
   "bin": {
-    "mpd2fm": "./bin/daemon.js"
+    "mpd2fm": "./bin/cli.js"
   },
   "devDependencies": {},
   "scripts": {
